requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
click==8.3.0
colorama==0.4.6
distro==1.9.0
fastapi==0.121.0
groq==0.33.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pydantic==2.12.4
pydantic-core==2.41.5
python-dotenv==1.2.1
sniffio==1.3.1
starlette==0.49.3
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.38.0
requests==2.32.3
tabulate==0.9.0
Jinja2==3.1.4
cryptography==43.0.3
authlib==1.3.0
sqlmodel==0.0.22
aiosqlite==0.20.0
loguru==0.7.2
pytest==8.3.3
respx==0.20.2
orjson==3.10.7
transformers==4.44.2
sqlmodel>=0.0.16
